+++ packages/ui/src/index.ts
@@ -1,4 +1,2 @@
-export * from './utils/cn';
-
 export { default as i18n } from './services/i18n';
 export * from './stores/notifications-store';
