+++ packages/ui/src/legacy/components/DownloadButton/DownloadButton.tsx
@@ -1,6 +1,6 @@
 import { ArrowDownTrayIcon } from '@heroicons/react/24/outline';
 
-import { cn } from '@/utils/cn';
+import { cn } from '@/utils';
 
 export type DownloadButtonProps = {
   className?: string;
