+++ packages/ui/tsup.config.js
@@ -13,7 +13,8 @@ export default defineConfig({
     path.resolve(__dirname, 'src/index.ts'),
     path.resolve(__dirname, 'src/components/index.ts'),
     path.resolve(__dirname, 'src/hooks/index.ts'),
-    path.resolve(__dirname, 'src/legacy/index.ts')
+    path.resolve(__dirname, 'src/legacy/index.ts'),
+    path.resolve(__dirname, 'src/utils.ts')
   ],
   format: 'esm',
   outDir: path.resolve(__dirname, 'dist'),
